Rework UAC/MIDI
* Split UAC from MIDI
* Split UAC into UAC1, UAC2, UAC3 as descriptors and
  other constants are often defined differently
