+++ usb_protocol/emitters/descriptors/uac2.py
@@ -6,11 +6,8 @@
 from contextlib import contextmanager
 
 from .. import emitter_for_format
-from ...types.descriptors.uac  import *
 from ...types.descriptors.uac2 import *
 from ...emitters.descriptor    import ComplexDescriptorEmitter
-
-###################### Audio #########################
 
 # Create our emitters.
 InterfaceAssociationDescriptorEmitter          = emitter_for_format(InterfaceAssociationDescriptor)
